+++ packages/Angular/Generic/markdown/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @memberjunction/ng-markdown
 
+## 6.1.0-edge.4
+
+### Patch Changes
+
+- @memberjunction/markdown-core@6.1.0-edge.4
+
+## 6.1.0-edge.3
+
+### Patch Changes
+
+- @memberjunction/markdown-core@6.1.0-edge.3
+
 ## 6.1.0-edge.2
 
 ### Patch Changes